# Use `CMOV` instruction to speed up `min()`