This increases performance of `min()` by 1-18% depending on scenario.

```
                              │   old.txt    │               new.txt               │
                              │    sec/op    │   sec/op     vs base                │
LevenshteinDistance-8            61.08n ± 0%   56.71n ± 0%   -7.15% (p=0.000 n=10)
LevenshteinDistanceBigLate-8    11.239µ ± 0%   9.230µ ± 0%  -17.88% (p=0.000 n=10)
LevenshteinDistanceBigEarly-8    9.424µ ± 0%   9.295µ ± 0%   -1.37% (p=0.000 n=10)
geomean                          1.863µ        1.694µ        -9.06%
```

Fixes #50